Creating a new branch register_middlwares_gates

+++ server/app/Http/Controllers/api/EnseignantController.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Controllers\api;
+use App\Events\storeuser;
 use App\Http\Controllers\AuthController;
 use App\Http\Resources\EnseignantResource;
 
@@ -14,6 +15,7 @@ use App\Models\Enseignant;
 
 use App\Traits\HttpResponses;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class EnseignantController extends Controller
 {
@@ -21,7 +23,8 @@ class EnseignantController extends Controller
      // A class that handles the success and error messages
      use HttpResponses;
 //=========================================================== The access is retricted for:AdminUae|President ==============================================
-        
+
+    
  
      /**
      * Index() it's a methode that serve to display all the profs with there etablissement and grade.
@@ -32,9 +35,10 @@ class EnseignantController extends Controller
        public function index()
 
           {   
-                       
-                 return EnseignantResource::collection(Enseignant::with('etablissement','grade')->latest()->paginate(10)); 
-                                
+            if (Gate::allows('check_role', [0])) {          
+           return EnseignantResource::collection(Enseignant::with('etablissement','grade')->latest()->paginate(10)); 
+           }
+            return $this->error('','ACCES INTERDIT ',403);                    
           }
 
 //============================================================== The access is retricted for:AdminUEtab ========================================================          
@@ -51,25 +55,35 @@ class EnseignantController extends Controller
 
           {  
 
+              
+              if (Gate::allows('check_role', [1])) { 
                  $enseignant=new Enseignant();
                  $grade_id=$enseignant->IdGrade($request['Grade']);                                       //IdGrade() it's a method that return the id of the grade 
-                 $etablissement_id=1;//auth()->user()->administrateur->etablissement_id                  //the security developper should approuve it
+                                
                  $enseignant->PPR = $request['PPR'];
                  $enseignant->nom = $request['nom'];
                  $enseignant->prenom = $request['prenom'];
                  $enseignant->date_naissance = $request['DateNaissance'];
-                 $enseignant->etablissement_id = $etablissement_id;
-                 $enseignant->grade_id =$grade_id;  //
-                 // $enseignant->user_id = $request['IdUser'];                                           // i talked with the postgres admin to add a trigger for this  ;
+                 $enseignant->etablissement_id = auth()->user()->administrateur->etablissement_id  ;
+                 $enseignant->grade_id =$grade_id;  
+                                                 
                  $enseignant->email_perso=$request['email_perso'];
                  $enseignant->save();
-                 return $this->succes("","added successfully");
+                
+                 $id=event (new storeuser($request->input('email_perso'),4,$request->input('nom'),$request->input('prenom')));
+                 $enseignant->user_id = $id[0];
+                 $enseignant->save();
+
+
+                 return $this->succes("","added successfully");}
+                 return $this->error('','ACCES INTERDIT ',403);
+
                  // return new EnseignantResource(Enseignant::create($request->all()));it's another method but we should know the etab_id 
 
           }
 //======================================================= The access is retricted for:AdminUEtab|AdminUae|DireteurEtab|President ==================================================
      
-    
+    //the access to this method is restricted for 0,3 and also 1 and 2
      /**
      * Show() this method serve to display a specified teacher with the name and the city of his etablissement and also his  grade  .
      * @param  int  $id it's IDEnseignant!!!!!!
@@ -78,8 +92,12 @@ class EnseignantController extends Controller
    
        public function show($id)
 
-          {                                    
-                 return new EnseignantResource(Enseignant::with('etablissement','grade')->FindOrFail($id));                    
+          {      if (Gate::allows('check_role', [0,3]) || Gate::allows('direct_ens',$id) || Gate::allows('admin_ens',$id) ) {  
+
+                 return new EnseignantResource(Enseignant::with('etablissement','grade')->FindOrFail($id));
+                }
+                 
+                return $this->error('','ACCES INTERDIT ',403);                   
                  
           }
         
@@ -96,7 +114,7 @@ class EnseignantController extends Controller
 
        public function update( UpdateEnseignantRequest $request, $id)
 
-          {               
+          {       if ( Gate::allows('admin_ens',$id)) {             
                       
                  $enseignant=Enseignant::find($id);
                  $grade_id=$enseignant->IdGrade($request['Grade']);  
@@ -108,6 +126,8 @@ class EnseignantController extends Controller
                  $enseignant->email_perso=$request['email_perso'];
                  $enseignant->save();
                  return $this->succes("","updated successfully");
+                }
+                 return $this->error('','ACCES INTERDIT ',403);  
                               
           }
 //======================================= The access is retricted for:AdminEtab ==============================================
@@ -121,11 +141,13 @@ class EnseignantController extends Controller
 
        public function destroy($id)
 
-          {                
+          {       
+              if ( Gate::allows('admin_ens',$id) ) {           
                  $ens= Enseignant::FindOrfail($id);              
                  unlink(public_path('uploads').'/'.$ens->image);                                         //destroy the appropriate image .       
                  $ens->delete();              
-                 return $this->succes("","enseignant deleted successfully");  
+                 return $this->succes("","enseignant deleted successfully"); }
+                 return $this->error('','ACCES INTERDIT ',403);  
           }
 
 //=============================================== The access is retricted for:Enseignant ===========================================================================                
@@ -138,10 +160,11 @@ class EnseignantController extends Controller
      */
 
        public function ShowMyInterventions($id)
-          {                
+          {      if (auth()->user()->role==4 && auth()->user()->id==$id ){          
                 
                                 //this method display all the interventions of a specified prof
-                               return new EnseignantInterventionResource(Enseignant::with('interventions.etablissement')->find($id));
+                               return new EnseignantInterventionResource(Enseignant::with('interventions.etablissement')->find($id));}
+                               return $this->error('','ACCES INTERDIT ',403);  
                                 
                         
                 }
@@ -159,7 +182,12 @@ class EnseignantController extends Controller
      */
           public function ShowMyPayments($id)
           {
-                     return EnseignantPaymentsResource::collection(Enseignant::where('user_id',$id)->with('paiements.etablissement'));
+            if (auth()->user()->role==4 && auth()->user()->id==$id ){
+               return EnseignantPaymentsResource::collection(Enseignant::where('user_id',$id)->with('paiements.etablissement'));
+            }
+            return $this->error('','ACCES INTERDIT ',403);  
+
+                     
        
           }
    
@@ -177,6 +205,10 @@ class EnseignantController extends Controller
        public function UploadMyImage( Request $request,$id)
        
           {
+              if (Gate::allows('ens_check_id', $id)) {
+
+              
+
 
                  $request->validate([ 'image'=>'required|max:1024|mimes:png,jpg,png' ]);
 
@@ -198,6 +230,9 @@ class EnseignantController extends Controller
                  {
                  return $this->succes("","image uploaded successfully");    
                  }
+              }
+              return $this->error('','ACCES INTERDIT ',403); 
+
           }
 //==================================== The access is retricted for:Enseignant   ============================================================                
            
@@ -210,7 +245,12 @@ class EnseignantController extends Controller
 
        public function ShowMyProfil($id)
           {
-                 return new EnseignantResource(Enseignant::where('user_id',$id)->with('etablissement','grade')->first());
+              if (Gate::allows('ens_check_id', $id)) {
+                     return new EnseignantResource(Enseignant::where('user_id',$id)->with('etablissement','grade')->first());
+              }
+              return $this->error('','ACCES INTERDIT ',403); 
+              
+                
           }
 
 //======================================= The access is retricted for:Enseignant ======================================= 
@@ -223,9 +263,11 @@ class EnseignantController extends Controller
      */
        public function MyHours ($id)
 
-          {
+          {    if (Gate::allows('check_role', [4])) {    
                  $enseignant=new Enseignant();
-                 return response()->json($enseignant->Hours($id));
+                 return response()->json($enseignant->Hours($id));}
+                 return $this->error('','ACCES INTERDIT ',403); 
+
           }
 
 //==================================== The access is retricted for:Enseignant ============================================
@@ -241,11 +283,15 @@ class EnseignantController extends Controller
        public function UpdateMyEmail( UpdateEnseignantRequest $request ,$id)
 
           {
+              if (Gate::allows('ens_check_id', $id)) {
                  $enseignant=Enseignant::where('user_id',$id)->first();  
                  $enseignant->email_perso=$request['email_perso'];   
                  $enseignant->save();  
                  return $this->succes("","email updated successfully");
           }
+          return $this->error('','ACCES INTERDIT ',403); 
+
+       }
 
                 
 //
