+++ server/app/Http/Controllers/api/GradeController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\UpdateGradeRequest;
 use App\Http\Resources\GradeResource;
 use App\Models\Grade;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class GradeController extends Controller
 {
@@ -23,7 +24,11 @@ class GradeController extends Controller
 
        public function index()
           {
+              if (Gate::allows('check_role', [0,3,1])) { 
+
                  return GradeResource::collection(Grade::latest()->paginate(10));
+              }
+              return $this->error('','ACCES INTERDIT ',403);
           }
 
 
@@ -37,12 +42,15 @@ class GradeController extends Controller
      */
        public function store(StoreGradeRequest $request)
           {
-            
+              
+              if (Gate::allows('check_role', [0,1])) { 
                  $grade= new GradeResource(Grade::create($request->all()));
                  if($grade)
                  {
                  return response()->json(["message"=>"added successfuly"]);
                  }
+              }
+              return $this->error('','ACCES INTERDIT ',403);
           }
 
 
@@ -56,8 +64,11 @@ class GradeController extends Controller
      */
        public function show($id)
           {   
+              if (Gate::allows('check_role', [0,1])) { 
             
                  return new GradeResource(Grade::findOrFail($id));
+              }
+              return $this->error('','ACCES INTERDIT ',403);
         
           }
 
@@ -73,8 +84,11 @@ class GradeController extends Controller
      */
        public function update(UpdateGradeRequest $request, $id)
           {
+              if (Gate::allows('check_role', [0,1])) { 
                  Grade::findOrFail($id)->update($request->all());
                  return response()->json(["message"=>" Updated successfully"]);
+              }
+              return $this->error('','ACCES INTERDIT ',403);
           }
 
 
@@ -88,12 +102,15 @@ class GradeController extends Controller
      */
        public function destroy($id)
           {
+              if (Gate::allows('check_role', [0,1])) { 
                  $grade=Grade::FindOrFail($id);
                  if($grade)
                  {
                  $grade->delete();
                  return response()->json(["message"=>"deleted successfuly"]);
                  }
+              }
+              return $this->error('','ACCES INTERDIT ',403);
           }
 }
