+++ server/app/Http/Controllers/api/InterventionController.php
@@ -11,6 +11,7 @@ use App\Http\Resources\InterventionResource;
 use App\Http\Resources\InterventionShowMoreResource;
 use Illuminate\Support\Facades\DB;
 use App\Traits\HttpResponses;
+use Illuminate\Support\Facades\Gate;
 
 class InterventionController extends Controller
 {
@@ -28,7 +29,12 @@ class InterventionController extends Controller
      */
     public function index()
     {
-        //return EnseignantResource::collection(Enseignant::with('etablissement','grade')->latest()->paginate(10)); 
+       if (Gate::allows('check_role', [1])) {  
+        //return EnseignantResource::collection(Enseignant::with('etablissement','grade')->latest()->paginate(10));
+        }
+        return $this->error('','ACCES INTERDIT ',403);
+
+
     }
 
 //===========================================  The access is retricted for:AdminEtab ================================================== 
@@ -44,10 +50,11 @@ class InterventionController extends Controller
        public function store(StoreInterventionRequest $request)
 
           {
+              if (Gate::allows('check_role', [1])) {  
               
                  $intervention=new Intervention();
                  $enseignant=$intervention->IdEnseignant($request['PPR']);
-                 $etablissement=1;     //auth()->user()->administrateur->etablissement_id;
+                 $etablissement=auth()->user()->administrateur->etablissement_id;
                  $intervention-> intitule_intervention= $request['IntituleIntervention'];
                  $intervention-> annee_univ= $request['AnneeUniv'];
                  $intervention-> semestre= $request['Semestre'];
@@ -58,6 +65,8 @@ class InterventionController extends Controller
                  $intervention->etablissement_id= $etablissement ;
                  $intervention->save();
                  return $this->succes("","Intervention added successfully");
+              }
+              return $this->error('','ACCES INTERDIT ',403);
 
                 
           }
@@ -75,7 +84,10 @@ class InterventionController extends Controller
        public function show($id)
 
           {
+              if (Gate::allows('check_role', [1,3])) {  
                  return new InterventionResource(Intervention::with('enseignant','etablissement')->findOrFail($id));
+              }
+              return $this->error('','ACCES INTERDIT ',403);
 
           }
 
@@ -93,7 +105,7 @@ class InterventionController extends Controller
        public function update(UpdateInterventionRequest $request, $id)
 
   
-          {                
+          {          if (Gate::allows('check_role', [1])) {         
                  $intervention=Intervention::FindOrFail($id);   
                  $enseignant=$intervention->IdEnseignant($request['PPR']);                                //it's a method that return the id of the enseignant 
                  // $etablissement= 1; //auth()->user()->administrateur->etablissement_id; 
@@ -107,6 +119,8 @@ class InterventionController extends Controller
                  //$intervention->etablissement_id=$etablissement;                                        // we will not use it because when he update the etablissement_id remain the same 
                  $intervention->save();
                  return $this->succes("","Intervention updated successfully");
+              }
+              return $this->error('','ACCES INTERDIT ',403);
           }
 
 
@@ -123,12 +137,15 @@ class InterventionController extends Controller
        public function destroy($id)
 
           {
+              if (Gate::allows('check_role', [1])) {    
                  $intervention= Intervention::find($id);
                  if($intervention)
                  {
                  $intervention->delete();
                  return $this->succes("","intervention deleted successfully");
                  }
+              }
+              return $this->error('','ACCES INTERDIT ',403);
 
           }
 
@@ -149,7 +166,10 @@ class InterventionController extends Controller
        public function ShowMore($id)
 
           {    
+              if (Gate::allows('check_role', [1,0,2,3])) {  
                  return new InterventionShowMoreResource(Intervention::with('enseignant','etablissement')->find($id));
+              }
+              return $this->error('','ACCES INTERDIT ',403);
           }
 
 
@@ -167,9 +187,12 @@ class InterventionController extends Controller
 
        public function activeVisaUae(UpdateInterventionRequest $request,$id)
           {
+              if (Gate::allows('check_role', [3])) {  
                  $intervention =Intervention::find($id);
                  $intervention->visa_uae=$request['VisaUae'];
                  $intervention->save();
+              }
+              return $this->error('','ACCES INTERDIT ',403);
           }
 
 
@@ -185,9 +208,12 @@ class InterventionController extends Controller
 
        public function activeVisaEtab(UpdateInterventionRequest $request,$id)
           {
+              if (Gate::allows('check_role', [2])) {  
                  $intervention =Intervention::find($id);
                  $intervention->visa_etab=$request['VisaEtab'];
                  $intervention->save();
+              }
+              return $this->error('','ACCES INTERDIT ',403);
           }
 
 
@@ -208,21 +234,24 @@ class InterventionController extends Controller
 
         
           {  
+              if (Gate::allows('check_role', [1,2])) { 
            
-                /* should be approved By the security developper */
+                
 
-                //    $role=auth()->user()->role;
-                //    if($role==1)
-                //    {
-                //     $etab_id=auth()->user()->administrateur->etablissement_id;
-                //    }
-                //    elseif($role==2)
-                //    {
-                //     $etab_id=auth()->user()->directeur->etablissement_id;
+                   $role=auth()->user()->role;
+                   if($role==1)
+                   {
+                   $etab_id=auth()->user()->administrateur->etablissement_id;
+                   }
+                    elseif($role==2)
+                   {
+                     $etab_id=auth()->user()->directeur->etablissement_id;
                         
-                //    }
-                 $etab_id=1;
+                    }
+                
                  return  InterventionResource::collection(Intervention::where('etablissement_id',$etab_id)->with('enseignant','etablissement')->latest()->paginate(10));
+              }
+              return $this->error('','ACCES INTERDIT ',403);
            
 
 
@@ -239,8 +268,11 @@ class InterventionController extends Controller
        
        public function EnseignantInterventions($id)
           {
+              if (Gate::allows('check_role', [3,0,1,2])) { 
 
                  return  InterventionResource::collection(Intervention::where('enseignant_id',$id)->with('enseignant','etablissement')->latest()->paginate(5));
+              }
+              return $this->error('','ACCES INTERDIT ',403);
 
           }
